drop debug hooks; rename to index_ownership
