+++ python/sglang/srt/mem_cache/memory_pool.py
@@ -58,13 +58,8 @@ from sglang.srt.layers.quantization.fp4_kv_cache_quant_method import (
 )
 from sglang.srt.layers.radix_attention import RadixAttention
 from sglang.srt.mem_cache.allocator.mamba import MambaSlotAllocator
-from sglang.srt.mem_cache.index_alias_guard import (
-    install_debug_hooks as install_index_alias_hooks,
-)
-from sglang.srt.mem_cache.index_alias_guard import (
-    register_req_to_token,
-)
 from sglang.srt.mem_cache.index_key_cache import IndexKeyCache
+from sglang.srt.mem_cache.index_ownership import register_req_to_token
 from sglang.srt.mem_cache.kv_vmm_backing import KvVmmBufferOwner
 from sglang.srt.mem_cache.layout.page_major import (
     build_page_major_mamba_views,
@@ -288,10 +283,8 @@ class ReqToTokenPool:
         self.free_slots = list(range(1, self._alloc_size))
         self.req_generation = torch.zeros(self._alloc_size, dtype=torch.int64)
 
-        # Unconditional: the allocators consult this on every deferred free.
+        # The allocators consult this on every deferred free.
         register_req_to_token(self.req_to_token)
-        if envs.SGLANG_DEBUG_MEMORY_POOL.get():
-            install_index_alias_hooks()
 
     def write(self, indices, values):
         self.req_to_token[indices] = values
